+++ app/javascript/utilities/utilities.ts
@@ -284,4 +284,4 @@ export const critHitDiff = 8
 export const critMorale = 4
 export const titleName = "A Hex Too Far"
 export const subtitleName = "Light Tactical Battle System"
-export const serverVersion = "0.57"
+export const serverVersion = "0.58"